+++ models/role/schema.js
@@ -16,9 +16,27 @@ module.exports = new mongoose.Schema({
         maxLength: 20,
         validate: snowflakeValidator,
     },
+    channel: {
+        type: String,
+        required: true,
+        maxLength: 20,
+        validate: snowflakeValidator,
+    },
+    message: {
+        type: String,
+        required: true,
+        maxLength: 20,
+        validate: snowflakeValidator,
+    },
     emoji_name: {
         type: String,
         required: true,
+    },
+    role: {
+        type: String,
+        required: false, // make this required once the transition happens
+        maxLength: 20,
+        validate: snowflakeValidator,
     },
     role_name: {
         type: String,