+++ models/settings/settings.js
@@ -18,14 +18,6 @@ class Settings {
         return this._model.guild;
     }
 
-    get roleMessage() {
-        return this._model.role_watch_message;
-    }
-
-    get welcomeChannel() {
-        return this._model.welcome_channel;
-    }
-
     get auditChannel() {
         return this._model.audit_channel;
     }
@@ -87,20 +79,6 @@ class Settings {
     setGuild(guild) {
         if (typeof guild === 'string') {
             this._model.guild = guild;
-        }
-        return this;
-    }
-
-    setRoleMessage(messageID) {
-        if (typeof messageID === 'string') {
-            this._model.role_watch_message = messageID;
-        }
-        return this;
-    }
-
-    setWelcomeChannel(channel) {
-        if (typeof channel === 'string') {
-            this._model.welcome_channel = channel;
         }
         return this;
     }
@@ -243,8 +221,6 @@ class Settings {
     toAPI() {
         return {
             guild: this._model.guild,
-            roleWatchMessage: this._model.role_watch_message,
-            roleWatchChannel: this._model.welcome_channel,
             musicWatchChannel: this._model.music_channel,
             auditChannel: this._model.audit_channel,
             commandPrefix: this._model.command_prefix,