+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added support for the bot to be on multiple discord servers
 - added support for bot admin commands to configure the bot settings
 - added the ability to automatically parse the role reaction emojis
+- added the ability to watch multiple role reaction messages
+- added the ability to update a role reaction message
+- added the ability to list all the role reaction messages that botomir is currently  watching
+- added the ability to remove a role reaction message without just updating it
 - added a bunch of new commands
 - added moderation tools to keep message logs
 - added the ability to handle webhooks from github and gitlab
@@ -17,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added scheduling recurring messages using crontab formatted strings
 - added scheduling reminder messages to be sent
 - added listing the scheduled recurring messages
+- added removing a scheduled recurring message
 - can automatically render message links 
 - can automatically assign roles based on reactions to a message
 - can generate Spotify playlists